requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.0.0
scikit-learn>=1.2.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
